Refactor ldClient into a separate file. Pass ldClient through broadcast. Rename lib/launchDarkly to utils
