+++ src/components/LaunchDarkly.js
@@ -2,6 +2,7 @@
 import React from "react";
 import { Broadcast } from "react-broadcast";
 
+import LDClient from "../lib/LDClient";
 import { BROADCAST_CHANNEL } from "../constants/LaunchDarkly.js";
 import { UserType } from "../types/User";
 
@@ -13,15 +14,11 @@ type Props = {
 
 export default function LaunchDarkly (props:Props) {
   const { clientId, user, children } = props;
-  const value = {
-    clientId,
-    user
-  };
+  const ldClient = LDClient.init(clientId, user);
 
   return (
-    <Broadcast channel={BROADCAST_CHANNEL} value={value}>
+    <Broadcast channel={BROADCAST_CHANNEL} value={ldClient}>
       {children}
     </Broadcast>
   );
 }
-
