+++ src/index.js
@@ -1,3 +1,3 @@
 export LaunchDarkly from "./components/LaunchDarkly";
 export FeatureFlag from "./components/FeatureFlag";
-export { getAllFeatureFlags } from "./lib/launchDarkly";
+export { getAllFeatureFlags } from "./lib/utils";
